requirements.txt
numpy==1.26.0
opencv-python==4.8.0.76
matplotlib==3.7.2
easyocr==1.7.1
pillow
google-generativeai
flask
flask_cors
typing_extensions
inference-sdk